linux syscalls: cast uint64_t sizes to size_t for 32-bit Windows
memory.cpp: page_align_up returns uint64_t but the memory manager API
takes size_t; cast aligned_length/aligned_old_size/aligned_new_size/
clear_size at the assignment so the rest of the function stays warning
free.

file.cpp: iov_len is read as a 64-bit guest field but used as buffer
size / fread count / string_view length, all of which are size_t. Cast
at each use site.

+++ src/linux-emulator/syscalls/file.cpp
@@ -940,12 +940,12 @@ void sys_writev(const linux_syscall_context& c)
             continue;
         }
 
-        std::vector<uint8_t> buffer(iov_len);
-        c.emu.read_memory(iov_base, buffer.data(), iov_len);
+        std::vector<uint8_t> buffer(static_cast<size_t>(iov_len));
+        c.emu.read_memory(iov_base, buffer.data(), static_cast<size_t>(iov_len));
 
         if (fd == 1 || fd == 2)
         {
-            const auto sv = std::string_view(reinterpret_cast<const char*>(buffer.data()), iov_len);
+            const auto sv = std::string_view(reinterpret_cast<const char*>(buffer.data()), static_cast<size_t>(iov_len));
             auto& callback = (fd == 1) ? c.emu_ref.on_stdout : c.emu_ref.on_stderr;
 
             if (callback)
@@ -955,7 +955,7 @@ void sys_writev(const linux_syscall_context& c)
             else
             {
                 auto* target = (fd == 1) ? stdout : stderr;
-                fwrite(buffer.data(), 1, iov_len, target);
+                fwrite(buffer.data(), 1, static_cast<size_t>(iov_len), target);
                 fflush(target);
             }
 
@@ -963,7 +963,7 @@ void sys_writev(const linux_syscall_context& c)
         }
         else if (fd_entry->handle)
         {
-            const auto written = fwrite(buffer.data(), 1, iov_len, fd_entry->handle);
+            const auto written = fwrite(buffer.data(), 1, static_cast<size_t>(iov_len), fd_entry->handle);
             total_written += static_cast<int64_t>(written);
         }
         else
@@ -1806,8 +1806,8 @@ void sys_readv(const linux_syscall_context& c)
             return;
         }
 
-        std::vector<uint8_t> buffer(iov_len);
-        const auto bytes_read = fread(buffer.data(), 1, iov_len, fd_entry->handle);
+        std::vector<uint8_t> buffer(static_cast<size_t>(iov_len));
+        const auto bytes_read = fread(buffer.data(), 1, static_cast<size_t>(iov_len), fd_entry->handle);
 
         if (bytes_read > 0)
         {
